+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,10 @@
         android:theme="@style/Theme.DemoApp"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SecondActivity"
+            android:name=".mvp.ui.SecondActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".mvvm.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
