+++ app/src/main/java/com/example/demoapp/mvp/model/SecondModel.kt
@@ -1,18 +1,18 @@
-package com.example.demoapp.repositories
+package com.example.demoapp.mvp.model
 
 import com.example.demoapp.UserList
-import com.example.demoapp.UserList.User
+import com.example.demoapp.mvp.contracts.SecondActivityContract
 import com.example.demoapp.data.remote.ReqresApi
 import com.example.demoapp.other.Resource
 import com.example.demoapp.other.Status
+import com.example.demoapp.repositories.DataStoreRepository
 import kotlinx.coroutines.flow.Flow
 
-class DefaultUserRepository(
+class SecondModel(
     private val userDataStoreRepository: DataStoreRepository,
     private val reqresApi: ReqresApi
-): UserRepository {
-
-    override suspend fun deleteUser(user: User) {
+): SecondActivityContract.Model {
+    override suspend fun deleteUser(user: UserList.User) {
         userDataStoreRepository.deleteData(user)
     }
 
@@ -39,7 +39,7 @@ class DefaultUserRepository(
             Resource.error("Couldn't reach the server. Check your internet connection", null)
         }
 
-        if (result.status==Status.SUCCESS){
+        if (result.status== Status.SUCCESS){
             result.data?.userData?.forEach {
                 userDataStoreRepository.addData(it)
             }
