+++ app/src/main/res/layout/activity_second.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#EEFDFB"
-    tools:context=".ui.SecondActivity">
+    tools:context=".mvp.ui.SecondActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvUserSecond"
